refactor : move pin static mehotd to PinUtil class
