+++ src/main/java/bowling/common/StaticVariables.java
@@ -1,0 +1,4 @@
+package bowling.common;
+
+public class StaticVariables {
+}
