+++ src/main/java/bowling/domain/Pin.java
@@ -43,32 +43,6 @@ public class Pin {
                 .orElse(false);
     }
 
-    public static int getNumOfPin(Pin pin) {
-        return Optional.ofNullable(pin)
-                .map(Pin::getNumOfFellPin)
-                .orElse(-1);
-    }
-
-    public static String toView(Pin pin) {
-        return Optional.ofNullable(pin)
-                .map(Pin::toString)
-                .orElse(" ");
-    }
-
-    public static String toView(Pin firstTry, Pin secondTry) {
-        return Optional.ofNullable(firstTry)
-                .filter(first -> first.isSpare(secondTry))
-                .map(fellPin -> "/")
-                .orElse(toView(secondTry));
-    }
-
-    public static String toSplitor(Pin pin) {
-        return Optional.ofNullable(pin)
-                .filter(fellpin -> !fellpin.toString().equals(STRIKE))
-                .map(fellPin -> "|")
-                .orElse(" ");
-    }
-
     @Override
     public String toString() {
         if(fellPin == 0 || fellPin == 10)
